% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Calculation.R
\name{Cal_HFD}
\alias{Cal_HFD}
\title{Calculate sap flow density using the HFD approach}
\usage{
Cal_HFD(
  data,
  T_up = NULL,
  T_low = NULL,
  T_side = NULL,
  Sym = NULL,
  Asym = NULL,
  K,
  L,
  Dst = 0.0025,
  Zax = 1.5,
  Ztg = 0.5
)
}
\arguments{
\item{data}{a data frame that includes the measured sap flow data}

\item{T_up}{column name for the temperature measured by the upper sensor}

\item{T_low}{column name for the temperature measured by the lower sensor}

\item{T_side}{column name for the temperature measured by the side sensor}

\item{Sym}{column name for the temperature difference between the upper and lower sensors}

\item{Asym}{column name for the temperature difference between the lower and side sensors}

\item{K}{a numeric value or a column name, K value indicates the dTasym value when zero sap flow occurs, see Nadezhdina et al. 2012}

\item{L}{a numeric value, sapwood depth, cm}

\item{Dst}{a numeric value, wood thermal diffusivity, cm2 s-1; default: 0.0025}

\item{Zax}{a numeric value, axial distance of the sensors, cm}

\item{Ztg}{a numeric value, tangential distance of the sensors, cm}
}
\value{
a data frame with an additional column of the calculated sap flow density
(SFD_hfd, g cm-2 h-1).
}
\description{
A function to calculate sap flow density using the HFD approach. Either the temperatures
that are directly measured by sensors (upper, lower and side) or temperature differences between the
sensors (dTsym and dTasym) must be provided.
}
\examples{
# load example data
df <- read.csv(file = system.file("extdata", "Soil_resp_example.csv", package = "FluxGapsR"),
header=T)

# convert the data into long format

}
