% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Calculation_HFD.R
\name{Cal_HFD}
\alias{Cal_HFD}
\title{Calculate sap flow density using the HFD approach}
\usage{
Cal_HFD(data, T_up, T_low, T_side, K, L, Dst = 0.0025, Zax = 1.5, Ztg = 0.5)
}
\arguments{
\item{data}{a data frame that includes the measured sap flow data}

\item{T_up}{column name for the temperature measured by the upper sensor}

\item{T_low}{column name for the temperature measured by the lower sensor}

\item{T_side}{column name for the temperature measured by the side sensor}

\item{K}{a numeric value (constant K) or a column name (dynamic K). K value indicates
the dTasym value when zero sap flow occurs, see Nadezhdina et al. 2012}

\item{L}{a numeric value; sapwood depth, cm}

\item{Dst}{a numeric value; wood thermal diffusivity, cm2 s-1; default: 0.0025}

\item{Zax}{a numeric value; axial distance of the sensors, cm. Default: 1.5}

\item{Ztg}{a numeric value; tangential distance of the sensors, cm. Default: 0.5}
}
\value{
a data frame with an additional column of the calculated sap flow density
(SFD_hfd, g cm-2 h-1).
}
\description{
A function to calculate sap flow density using the HFD approach (Nadezhdina et al. 2012).
The temperatures that are directly measured by sensors (upper, lower and side) must be provided.
}
\examples{
# load example data
df <- read.csv(file = system.file("extdata", "Soil_resp_example.csv", package = "FluxGapsR"),
header=T)

# convert the data into long format
df <- df \%>\%
  pivot_longer(cols = Temp1U:Temp8S,
               names_to = c(".value","Position"),
               names_sep = c(5)) \%>\%
  pivot_longer(cols = starts_with(c("K","Temp")),
               names_to = c(".value","Depth"),
               names_sep = -1) \%>\%
  pivot_wider(names_from = Position,
              values_from = Temp)

# calculate sap flow density using dynamic K values that are provided as a variable
df_hfd1 <- Cal_HFD(df,T_up = U,T_low = L,T_side = S,K=K,L=7.5)
tapply(df_hfd1$SFD_hfd, df_hfd1$Depth, summary)

# calculate sap flow density using constant K value
df <- df \%>\% select(-K)
df_hfd2 <- Cal_HFD(df,T_up = U,T_low = L,T_side = S,K=.6,L=7.5)
tapply(df_hfd2$SFD_hfd, df_hfd2$Depth, summary)

}
